requirements.txt
flask
tox
tox-travis
flask-wtf
python-dotenv
flask-sqlalchemy
flask-login
flask-bootstrap
